+++ sketch-compose-core/src/main/kotlin/com/github/panpf/sketch/compose/request/PainterImage.kt
@@ -4,10 +4,12 @@ import android.graphics.drawable.BitmapDrawable
 import androidx.compose.ui.graphics.asImageBitmap
 import androidx.compose.ui.graphics.painter.BitmapPainter
 import androidx.compose.ui.graphics.painter.Painter
+import com.github.panpf.sketch.cache.MemoryCache.Value
 import com.github.panpf.sketch.request.BitmapImage
 import com.github.panpf.sketch.request.ByteCountProvider
 import com.github.panpf.sketch.request.DrawableImage
 import com.github.panpf.sketch.request.Image
+import com.github.panpf.sketch.request.internal.RequestContext
 import com.google.accompanist.drawablepainter.DrawablePainter
 import kotlin.math.roundToInt
 
@@ -55,4 +57,9 @@ data class PainterImage(val painter: Painter, override val shareable: Boolean = 
 
             else -> 4 * width * height
         }
+
+    override fun cacheValue(
+        requestContext: RequestContext,
+        extras: Map<String, Any?>
+    ): Value? = null
 }